Remove all validation logic from SystemTestDefinition
SystemTestDefinition will now only define parameters for the test. 
Other classes will utilize these parameters when generating expected results and getting actual results
